+++ packages/go_router/lib/go_router.dart
@@ -7,6 +7,7 @@
 library go_router;
 
 export 'src/configuration.dart' show GoRouterState, GoRoute;
+export 'src/information_provider.dart' show BrowserState;
 export 'src/misc/extensions.dart';
 export 'src/misc/inherited_router.dart';
 export 'src/misc/refresh_stream.dart';